pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "parsit"
version = "0.1.0"
description = "Parsit VLM for Document Analysis"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.19.0",
    "transformers>=4.53.2",
    "accelerate>=1.0.0",
    "deepspeed>=0.15.0",
    "peft>=0.13.0",
    "bitsandbytes>=0.44.0",
    "flash-attn>=2.6.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "sentencepiece>=0.2.0",
    "protobuf>=4.25.0",
    "wandb>=0.17.0",
    "einops>=0.8.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "opencv-python>=4.9.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.4.0",
    "datasets>=2.20.0",
    "tokenizers>=0.20.0",
    "psutil>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

[tool.setuptools.packages.find]
include = ["parsit*"]